Move adapters.dataframe to adapters.table.
